Move status_testing.h into testutil
