+++ Firestore/core/test/firebase/firestore/nanopb/CMakeLists.txt
@@ -24,4 +24,5 @@ cc_test(
     GMock::GMock
     firebase_firestore_nanopb
     firebase_firestore_remote
+    firebase_firestore_testutil
 )
