+++ Firestore/core/test/firebase/firestore/remote/serializer_test.cc
@@ -53,8 +53,8 @@
 #include "Firestore/core/src/firebase/firestore/util/status.h"
 #include "Firestore/core/src/firebase/firestore/util/statusor.h"
 #include "Firestore/core/test/firebase/firestore/nanopb/nanopb_testing.h"
+#include "Firestore/core/test/firebase/firestore/testutil/status_testing.h"
 #include "Firestore/core/test/firebase/firestore/testutil/testutil.h"
-#include "Firestore/core/test/firebase/firestore/util/status_testing.h"
 #include "absl/strings/string_view.h"
 #include "absl/types/optional.h"
 #include "google/protobuf/stubs/common.h"
